Extract system tests
